Update README.md
E-commerce Concept App

Developed with React, Material UI, and Commerce JS.


Summary

The purpose of this project is to show conceptual styling and functionality for an e-commerce web application. The styling was done with the Material UI Framework. I took advantage of their grid system to create a clean and responsive interface. All the additional styling was added with the Material UI styles package. 

The Commerce JS API was used to serve as a backend for the product information and images. This API also provided additional functionality such as generating a  shopping cart and adding/removing products. 

You can start using this application by navigating to any product page and adding the item(s) to the cart. After doing so, click on the icon in the top right to view all the items in the shopping cart. Within this page, you can update quantity, delete a product, or empty the cart. There is also a checkout option that will lead you through a two-step process. Please note, this is only a simulation so no actual order will be placed and no real payment method will be accepted. After placing the order, the shopping cart will be cleared and the process can start over.